libxtrans fontconfig mesa xserver git update 28 Mar 2014
xserver          commit a2880699e8f1f576e1a48ebf25e8982463323f84
libxtrans        commit 68f60238c4224f954ff6556ae778c72e420175f0
fontconfig       commit fcba9ef01c978323fc71c17e455d3cd6ae35edcc
mesa             commit 029ccd773d01a5f801c809c499516d7b0c4cc3f8

+++ xorg-server/mi/mibitblt.c
@@ -730,7 +730,7 @@ miPutImage(DrawablePtr pDraw, GCPtr pGC, int depth,
         ChangeGC(NullClient, pGC, GCForeground | GCBackground, gcv);
         bytesPer = (long) h *BitmapBytePad(w + leftPad);
 
-        for (i = 1 << (depth - 1); i != 0; i >>= 1, pImage += bytesPer) {
+        for (i = (unsigned long) 1 << (depth - 1); i != 0; i >>= 1, pImage += bytesPer) {
             if (i & oldPlanemask) {
                 gcv[0].val = (XID) i;
                 ChangeGC(NullClient, pGC, GCPlaneMask, gcv);
